added framework for port connection
created a port to connect the popup and the background, so that the background gets a special message on port close, which supposedly i can pass state into, so that the background script then saves the popups state. no idea how to do the second bit, so for now the poup will store state every time the slider is set.
humph.