continuous value update
managed to get slider values to continuously update, noice.

now need to link the audio value to it.
method one, send a message, could be slow.
method two, link the audio thing to the popup, so can update the audio gain node directly in the popup, porbably better but we shall see.