can now delete nodes
when a tab is closed, the slider will now be removed from the extension.

to do this safely, a valid field has been added for each tab, and the id has changed from the stream id to the id of the tab.

Then when the tab is closed, if the tab id is currently registered as being monitered for audio, its valid field is set to false and a message is sent boradcasting the fact that this field is no longer valid.

as it turns up the popup closes when you close a tab anyway, and only generates sliders if the tab record is valid.