skeleton code for page
have started adding helper functions to the page document, for getting HTML5 audio from pages.
For now am going to skip the RTC connection stuff due to time, will be able to do that using my already good implementation in the background script anyway, dont need that code going to waste, so may as well use it.

next to get the code working with a gain node added in without errors