got the page access working! yes!
added an exta class to the html of the popup,
added some extra message passing logic to background and page, and added some to popup as well as long as a second map for holding page scripts, which now have a different object structure. I seem to constantly be passing in undefined things due to putting wrong parameters in place, but hopefully some code tidy up and renaming will sort that out.